Extract lerarenbeurs tests to seperate file
